apply related environment variables as config overrides
Using git-configuration to store overrides from the environment
is helpful as it more tighly integrates with the best configuration
system there is, and also is nicely visualizable.

It's now done for transport related configuration as well as all
other environment variables we previously queried on the fly.
