+++ git-repository/src/config/snapshot/_impls.rs
@@ -51,7 +51,7 @@ impl Deref for Snapshot<'_> {
     type Target = git_config::File<'static>;
 
     fn deref(&self) -> &Self::Target {
-        &self.plumbing()
+        self.plumbing()
     }
 }
