+++ git-repository/tests/fixtures/generated-archives/make_config_repos.tar.xz
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:3687c843a39942197ab8a31896b3a6c453295de7bb8d031642eb34b292a272da
-size 10816
+oid sha256:607f9cd2af225da2cff5640855d557b363949bd667454e772ae49a63deb21831
+size 13284
